fix: combine see detail and list feature by role
