--- book-store-management-backend/module/role/rolemodel/res_list_feature_by_role.go
@@ -1,8 +1,0 @@
-package rolemodel
-
-import "book-store-management-backend/module/feature/featuremodel"
-
-type ResListFeatureByRole struct {
-	// Data contains list of feature of role.
-	Data []featuremodel.ResFeatureDetail `json:"data"`
-}
