+++ book-store-management-backend/module/role/rolerepo/see_detail_role.go
@@ -1,9 +1,25 @@
 package rolerepo
 
 import (
+	"book-store-management-backend/module/feature/featuremodel"
 	"book-store-management-backend/module/role/rolemodel"
+	"book-store-management-backend/module/rolefeature/rolefeaturemodel"
 	"context"
+	"log"
 )
+
+type ListRoleFeaturesStore interface {
+	FindListFeatures(
+		ctx context.Context,
+		roleId string,
+	) ([]rolefeaturemodel.RoleFeature, error)
+}
+
+type ListAllFeatures interface {
+	ListFeature(
+		ctx context.Context,
+	) ([]featuremodel.Feature, error)
+}
 
 type FindRoleStore interface {
 	FindRole(
@@ -13,13 +29,19 @@ type FindRoleStore interface {
 }
 
 type seeRoleDetailRepo struct {
-	roleStore FindRoleStore
+	roleStore        FindRoleStore
+	roleFeatureStore ListRoleFeaturesStore
+	featureStore     ListAllFeatures
 }
 
 func NewSeeRoleDetailRepo(
-	roleStore FindRoleStore) *seeRoleDetailRepo {
+	roleStore FindRoleStore,
+	roleFeatureStore ListRoleFeaturesStore,
+	featureStore ListAllFeatures) *seeRoleDetailRepo {
 	return &seeRoleDetailRepo{
-		roleStore: roleStore,
+		roleStore:        roleStore,
+		roleFeatureStore: roleFeatureStore,
+		featureStore:     featureStore,
 	}
 }
 
@@ -32,9 +54,42 @@ func (biz *seeRoleDetailRepo) SeeRoleDetail(
 		return nil, errRole
 	}
 
+	features, errFeature := biz.featureStore.ListFeature(ctx)
+	if errFeature != nil {
+		return nil, errFeature
+	}
+	log.Print(features)
+
+	featuresRoleHas, errRoleFeature :=
+		biz.roleFeatureStore.FindListFeatures(ctx, roleId)
+	if errRoleFeature != nil {
+		return nil, errRoleFeature
+	}
+	log.Print(featuresRoleHas)
+
+	mapHasFeature := make(map[string]bool)
+	for _, v := range features {
+		mapHasFeature[v.Id] = false
+	}
+	for _, v := range featuresRoleHas {
+		mapHasFeature[v.FeatureId] = true
+	}
+
+	var featureDetails []featuremodel.ResFeatureDetail
+	for _, v := range features {
+		featureDetail := featuremodel.ResFeatureDetail{
+			Id:          v.Id,
+			Description: v.Description,
+			GroupName:   v.GroupName,
+			IsHas:       mapHasFeature[v.Id],
+		}
+		featureDetails = append(featureDetails, featureDetail)
+	}
+
 	var roleDetail rolemodel.ResSeeDetailRole
 	roleDetail.Id = role.Id
 	roleDetail.Name = role.Name
+	roleDetail.Data = featureDetails
 
 	return &roleDetail, nil
 }
