+++ book-store-management-backend/module/role/roletransport/ginrole/see_detail_role.go
@@ -4,9 +4,11 @@ import (
 	"book-store-management-backend/common"
 	"book-store-management-backend/component/appctx"
 	"book-store-management-backend/middleware"
+	"book-store-management-backend/module/feature/featurestore"
 	"book-store-management-backend/module/role/rolebiz"
 	"book-store-management-backend/module/role/rolerepo"
 	"book-store-management-backend/module/role/rolestore"
+	"book-store-management-backend/module/rolefeature/rolefeaturestore"
 	"github.com/gin-gonic/gin"
 	"net/http"
 )
@@ -25,9 +27,12 @@ func SeeDetailRole(appCtx appctx.AppContext) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		id := c.Param("id")
 
-		roleStore := rolestore.NewSQLStore(appCtx.GetMainDBConnection())
+		db := appCtx.GetMainDBConnection()
+		roleStore := rolestore.NewSQLStore(db)
+		roleFeatureStore := rolefeaturestore.NewSQLStore(db)
+		featureStore := featurestore.NewSQLStore(db)
 
-		repo := rolerepo.NewSeeRoleDetailRepo(roleStore)
+		repo := rolerepo.NewSeeRoleDetailRepo(roleStore, roleFeatureStore, featureStore)
 		requester := c.MustGet(common.CurrentUserStr).(middleware.Requester)
 
 		biz := rolebiz.NewSeeDetailRoleBiz(repo, requester)
